Add DiscardGenerator common interface
